+++ lib/simple_command_dispatcher.rb
@@ -9,13 +9,11 @@ require 'simple_command_dispatcher/services/command_service'
 require 'simple_command_dispatcher/version'
 
 module SimpleCommandDispatcher
-  # Provides a way to call SimpleCommands or your own custom commands in a more dymanic manner.
-  #
-  # For information about the simple_command gem, visit {https://rubygems.org/gems/simple_command}
+  # Provides a way to call your custom commands dynamically.
   #
   class << self
-    # Calls a *Command* given the command name, the modules the command belongs to
-    # and the parameters to pass to the command.
+    # Calls a *Command* given the command name, the namespace (modules) the command belongs to
+    # and the (request) parameters to pass to the command.
     #
     # @param command [Symbol, String] the name of the Command to call.
     #
@@ -25,18 +23,10 @@ module SimpleCommandDispatcher
     #    will both produce 'Api::AppName::V1', this string will be prepended to the command to form the Command
     #    to call (e.g. 'Api::AppName::V1::MySimpleCommand' = Api::AppName::V1::MySimpleCommand.call(*request_params)).
     #
-    # @param [Hash] options the options that determine how command and command_module are transformed.
-    # @option options [Boolean] :camelize (false) determines whether or not both class and module names should be
-    #    camelized.
-    # @option options [Boolean] :titleize (false) determines whether or not both class and module names should be
-    #    titleized.
-    # @option options [Boolean] :titleize_class (false) determines whether or not class names should be titleized.
-    # @option options [Boolean] :class_camelized (false) determines whether or not class names should be camelized.
-    # @option options [Boolean] :titleize_module (false) determines whether or not module names should be titleized.
-    # @option options [Boolean] :module_camelized (false) determines whether or not module names should be camelized.
-    #
-    # @param request_params [Array<Symbol>] the parameters to pass to the call method of the Command. This
-    #    parameter is simply passed through to the call method of the Command.
+    # @param request_params [Hash, Array, Object] the parameters to pass to the call method of the Command. This
+    #    parameter is simply passed through to the call method of the Command. Hash parameters are passed as
+    #    keyword arguments, Array parameters are passed as positional arguments, and other objects are passed
+    #    as a single argument.
     #
     # @return [Object] the Object returned as a result of calling the Command#call method.
     #
@@ -44,18 +34,20 @@ module SimpleCommandDispatcher
     #
     #  # Below call equates to the following:
     #  # Api::Carz4Rent::V1::Authenticate.call({ email: 'sam@gmail.com', password: 'AskM3!' })
-    #  SimpleCommandDispatcher.call(:Authenticate,
-    #     { api: :Api, app_name: :Carz4Rent, api_version: :V1 },
-    #     { email: 'sam@gmail.com', password: 'AskM3!' } ) # => Command result
+    #  SimpleCommandDispatcher.call(command: :Authenticate,
+    #     command_namespace: { api: :Api, app_name: :Carz4Rent, api_version: :V1 },
+    #     request_params: { email: 'sam@gmail.com', password: 'AskM3!' } ) # => Command result
     #
     #  # Below equates to the following: Api::Carz4Rent::V2::Authenticate.call('sam@gmail.com', 'AskM3!')
-    #  SimpleCommandDispatcher.call(:Authenticate,
-    #     ['Api', 'Carz4Rent', 'V2'], 'sam@gmail.com', 'AskM3!') # => Command result
+    #  SimpleCommandDispatcher.call(command: :Authenticate,
+    #     command_namespace: ['Api', 'Carz4Rent', 'V2'], 
+    #     request_params: ['sam@gmail.com', 'AskM3!']) # => Command result
     #
     #  # Below equates to the following:
     #  # Api::Auth::JazzMeUp::V1::Authenticate.call('jazz_me@gmail.com', 'JazzM3!')
-    #  SimpleCommandDispatcher.call(:Authenticate, ['Api::Auth::JazzMeUp', :V1],
-    #     'jazz_me@gmail.com', 'JazzM3!') # => Command result
+    #  SimpleCommandDispatcher.call(command: :Authenticate, 
+    #     command_namespace: ['Api::Auth::JazzMeUp', :V1],
+    #     request_params: ['jazz_me@gmail.com', 'JazzM3!']) # => Command result
     #
     def call(command:, command_namespace: {}, request_params: nil)
       # Create a constantized class from our command and command_namespace...