+++ lib/simple_command_dispatcher/errors/required_class_method_missing_error.rb
@@ -3,7 +3,10 @@
 module SimpleCommandDispatcher
   module Errors
     # This error is raised when a required class method is missing on the command class.
-    class Errors::RequiredClassMethodMissingError < StandardError
+    class RequiredClassMethodMissingError < StandardError
+      # Initializes a new RequiredClassMethodMissingError
+      #
+      # @param command_class_constant [Class] the command class that is missing the required method
       def initialize(command_class_constant)
         super("Class \"#{command_class_constant}\" does not respond_to? class method \"call\".")
       end